Use tags, instead of service name
When trying to determine which bound service for the target application
is the logstash service, use the tags on the service entity, not its
label.  This allows people to run multiple logstash instances in their
CloudFoundry (if they so desire), and use names other than 'logstash14'

Any services tagged 'logstash14' will be considered good enough for the
purposes of the auto-detection.